Moving wiki components to wiki folder
(cherry picked from commit 93bd1548d8abad00a8c22ec08a35c16bdac4813e)
